pipeline/requirements.txt
# Kraken
kraken
numpy<2.0
ipykernel

# Transformers
tiktoken
protobuf
pillow
transformers[torch,protobuf,tiktoken,sentencepiece]

pyspellchecker==0.8.1
